+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.55.0...HEAD)
 ------------------------------------------------------------------------
+### Changed
+- error codes
 
 [0.55.0](https://github.com/jshwi/docsig/releases/tag/v0.55.0) - 2024-06-12
 ------------------------------------------------------------------------